Refactor: Change folder structure, data handling
Replace the current project structure with a new one, which is optimized
for the setuptools install script. Change the data locations for
credentials, token and config as well for that purpose.

Add automatic change record clean up and implement error logging.

Handle the data location for the installed script.

**Attention**: The script doesn't work anymore without installing it as
the data locations are not valid, the only exception would be if you
explicitly place the required data into: /usr/local/data

Signed-off-by: Sebastian Fricke <sebastian.fricke.linux@gmail.com>