V0.2: cli.py: Update application CLI
Remove the logging feature as it was not very useful. Update the command
line interface options and add a more fine-grained configuration.

Update the application to use the gspread python module and include a
check for deleted and added items at plentymarkets.

Make use of a 'shared data' module, to reuse data pulled from the
Plentymarkets API, in multiple modules without polluting the parameters.

Signed-off-by: Sebastian Fricke <sebastian.fricke.linux@gmail.com>