V0.2: gsheet: Update the google sheet handling
Replace the self-made implementation with the gspread module in order to
make the code more readable and robust. The new module has one drawback
as it is not possible to update single values, it always writes the
whole file.
Add a new functionality to delete items, that were removed on
PlentyMarkets. As well as the functionality to add items, that are found
on PlentyMarkets but not on Facebook.

Signed-off-by: Sebastian Fricke <sebastian.fricke.linux@gmail.com>